+++ integration_tests/juniper_tests/src/codegen/derive_input_object.rs
@@ -1,4 +1,3 @@
-#[cfg(test)]
 use fnv::FnvHashMap;
 
 use juniper::DefaultScalarValue;